feat: move jwt signing to coinbase/x402
