+++ typescript/packages/coinbase-x402/package.json
@@ -38,6 +38,7 @@
     "vitest": "^3.0.5"
   },
   "dependencies": {
+    "@coinbase/cdp-sdk": "^1.1.1",
     "viem": "^2.23.1",
     "x402": "workspace:^",
     "zod": "^3.24.2"
