+++ typescript/packages/x402/package.json
@@ -38,7 +38,6 @@
     "vitest": "^3.0.5"
   },
   "dependencies": {
-    "@coinbase/cdp-sdk": "^1.1.1",
     "axios": "^1.7.9",
     "viem": "^2.23.1",
     "zod": "^3.24.2"
