+++ typescript/packages/x402/src/shared/index.ts
@@ -2,5 +2,4 @@ export * from "./json";
 export * from "./paywall";
 export * from "./base64";
 export * from "./network";
-export * from "./cdp";
 export * from "./middleware";
