--- typescript/packages/x402/src/version.ts
@@ -1,1 +1,0 @@
-export const version = "0.3.3";
